+++ web/lib/enums.ts
@@ -1,0 +1,9 @@
+// #region Cross platform API
+export enum PlatformEnum {
+  Capacitor = "capacitor",
+  Electron = "electron",
+  VSCode = "vscode",
+  Web = "web",
+  WebMobile = "web-mobile",
+}
+// #endregion